Update code to ES6 and Add time
