+++ index.html
@@ -5,7 +5,7 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="style.css">
-  <script defer src="index.js"></script>            
+  <script defer type="module" src="index.js"></script>        
   <title>Awesome Books</title>
 </head>
 <body>
@@ -19,6 +19,8 @@
       <li class="items contactBtn">Contact</li>
     </ul>
   </nav>
+  <section id="date" class="flex ">
+  </section>
   <section class="books">
     <h1>All awesome Books</h1>
     <div class="bookList"></div>
