+++ modules/get_local_storage.js
@@ -1,0 +1,5 @@
+export default () => {
+  const localdata = localStorage.getItem('localdata');
+  const dataStored = JSON.parse(localdata);
+  return dataStored;
+};
