+++ modules/update_local_storage.js
@@ -1,0 +1,4 @@
+export default (obj) => {
+  const localdata = JSON.stringify(obj);
+  localStorage.setItem('localdata', localdata);
+};
